# openapi.yaml
openapi: 3.0.0
info:
  title: 'GeoService API'
  version: 2.2.0
servers:
  -
    url: 'http://127.0.0.1:8085'
    description: 'Local environment'
paths:
...
  /api/v1/nationalities:
    get:
      tags:
        - Nationalities
      operationId: get_app_nationalities_api_v1_nationality_list
      parameters:
        -
          name: limit
          in: query
          description: 'Default 500'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: offset
          in: query
          description: 'Default 0'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: title
          in: query
          description: 'Nationality title'
          required: false
          content:
            application/json:
              schema:
                type: string
      responses:
        '200':
          description: 'List of nationalities.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NationalityResponse'
    post:
      tags:
        - Nationalities
      operationId: post_app_nationalities_api_v1_nationality_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNationalityActionRequest'
      responses:
        '200':
          description: 'Nationality created successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NationalityResponse'
  '/api/v1/nationalities/{id}':
    put:
      tags:
        - Nationalities
      operationId: put_app_nationalities_api_v1_nationality_update
      parameters:
        -
          name: id
          in: path
          description: 'Nationality uuid'
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNationalityActionRequest'
      responses:
        '200':
          description: 'Nationality updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NationalityResponse'
    delete:
      tags:
        - Nationalities
      operationId: delete_app_nationalities_api_v1_nationality_delete
      parameters:
        -
          name: id
          in: path
          description: 'Nationality uuid'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Nationality deleted successfully.'
  /api/v1/regions:
    get:
      tags:
        - Regions
      operationId: get_app_region_api_v1_region_list
      parameters:
        -
          name: limit
          in: query
          description: 'Default 500'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: offset
          in: query
          description: 'Default 0'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: title
          in: query
          description: 'Nationality title'
          required: false
          content:
            application/json:
              schema:
                type: string
      responses:
        '200':
          description: 'List of regions.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegionResponse'
    post:
      tags:
        - Regions
      operationId: post_app_region_api_v1_region_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRegionActionRequest'
      responses:
        '200':
          description: 'Region created successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionResponse'
  '/api/v1/regions/{id}':
    put:
      tags:
        - Regions
      operationId: put_app_region_api_v1_region_update
      parameters:
        -
          name: id
          in: path
          description: 'Region uuid'
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRegionActionRequest'
      responses:
        '200':
          description: 'Region updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionResponse'
    delete:
      tags:
        - Regions
      operationId: delete_app_region_api_v1_region_delete
      parameters:
        -
          name: id
          in: path
          description: 'Region uuid'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Region deleted successfully.'
...
  /api/v1/states:
    get:
      tags:
        - States
      operationId: get_app_state_api_v1_state_list
      parameters:
        -
          name: limit
          in: query
          description: 'Default 500'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: offset
          in: query
          description: 'Default 0'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: title
          in: query
          description: 'State title'
          required: false
          content:
            application/json:
              schema:
                type: string
      responses:
        '200':
          description: 'List of states.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StateResponse'
    post:
      tags:
        - States
      operationId: post_app_state_api_v1_state_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStateActionRequest'
      responses:
        '200':
          description: 'Nationality created successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateResponse'
  '/api/v1/states/{id}':
    put:
      tags:
        - States
      operationId: put_app_state_api_v1_state_update
      parameters:
        -
          name: id
          in: path
          description: 'State uuid'
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStateActionRequest'
      responses:
        '200':
          description: 'Nationality updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateResponse'
    delete:
      tags:
        - States
      operationId: delete_app_state_api_v1_state_delete
      parameters:
        -
          name: id
          in: path
          description: 'State uuid'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'State deleted successfully.'
...
components:
  schemas:
    NationalityResponse:
      required:
        - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
      type: object
    CreateNationalityActionRequest:
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 1
      type: object
    UpdateNationalityActionRequest:
      required:
        - title
      properties:
        title:
          type: string
      type: object
    RegionResponse:
      required:
        - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
      type: object
    CreateRegionActionRequest:
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 1
      type: object
    UpdateRegionActionRequest:
      required:
        - title
      properties:
        title:
          type: string
      type: object
    SubRegionResponse:
      required:
        - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        region:
          nullable: true
          oneOf:
            -
              $ref: '#/components/schemas/RegionResponse'
      type: object
...
    TimezoneResponse:
      required:
        - title
        - code
        - utc
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        code:
          type: string
        utc:
          type: string
      type: object
...
    CurrencyResponse:
      required:
        - name
        - code
        - symbol
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        symbol:
          type: string
      type: object
...
    CountryResponse:
      required:
        - title
        - iso2
        - iso3
        - numericCode
        - phoneCode
        - flag
        - tld
        - longitude
        - latitude
        - timezones
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        nativeTitle:
          type: string
          nullable: true
        iso2:
          type: string
        iso3:
          type: string
        numericCode:
          type: string
        phoneCode:
          type: string
        flag:
          type: string
        tld:
          type: string
        longitude:
          type: number
          format: float
        latitude:
          type: number
          format: float
        altitude:
          type: integer
          nullable: true
        currency:
          nullable: true
          oneOf:
            -
              $ref: '#/components/schemas/CurrencyResponse'
        subRegion:
          nullable: true
          oneOf:
            -
              $ref: '#/components/schemas/SubRegionResponse'
        timezones:
          type: array
          items:
            $ref: '#/components/schemas/TimezoneResponse'
      type: object
...
    StateResponse:
      required:
        - title
        - code
        - latitude
        - longitude
        - country
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        code:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        altitude:
          type: integer
          nullable: true
        type:
          type: string
          nullable: true
        country:
          $ref: '#/components/schemas/CountryResponse'
      type: object
    CreateStateActionRequest:
      required:
        - title
        - code
        - countryIso2
        - longitude
        - latitude
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 1
        code:
          type: string
          maxLength: 5
          minLength: 1
        countryIso2:
          type: string
          maxLength: 2
          minLength: 2
        longitude:
          type: number
          format: float
        latitude:
          type: number
          format: float
        altitude:
          type: integer
          nullable: true
        type:
          type: string
          maxLength: 50
          minLength: 1
          nullable: true
      type: object
    UpdateStateActionRequest:
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 1
          nullable: true
        type:
          type: string
          maxLength: 50
          minLength: 1
          nullable: true
        longitude:
          type: number
          format: float
          nullable: true
        latitude:
          type: number
          format: float
          nullable: true
        altitude:
          type: integer
          nullable: true
        countryIso2:
          type: string
          maxLength: 2
          minLength: 2
          nullable: true
      type: object
...
  securitySchemes:
    Bearer:
      type: http
      bearerFormat: JWT
      scheme: bearer
security:
  -
    Bearer: []